Moved and updated full screen modal example
